ui: receive button context menu
